core/requirements.txt
numpy==2.3.4
RPi.GPIO==0.7.1
spidev==3.8
